Refactor rotation/compression/backups routine of FileSink
Before the compression or the rotation, if the future file to be created is undifferentiable of the written file, then this file, after being closed, will be renamed by adding a unique identifier at the end of its name.
The compression takes place before the rotation, or when the logger closes if there is no rotation.
Rotation takes place when the corresponding function is positively evaluated.
The backups management takes place after the rotation or after compression or before closing.

Several elements must be taken into account to explain this choice of methodology:
- If we want the compression to be threaded, we must rename the file before compressing it to avoid potential conflict with the new file
- If we rename an already compressed file, its name will no longer accurately reflect the file it contains
- If we add a suffix to the compressed file, the extension will no longer make sense
- We also have to manage the compression extensions created by the user
- It avoids having to rename hundreds of potentially opened files in other programs or threads, we just rename the current log file once and for all at closing