Fix broken unit tests for Mypy
Happened due to Mypy upgrade from 0.961 to 0.971 (I don't know because
of which commit).

I wanted the Mypy tests to run even if Loguru wasn't intalled (just like
it does work with Pytest, using the local "loguru" folder only), but
it's not worth it because of the added complexity and because
configuring the development environment with "pip install -e .[dev]"
does install the package anyway.