Fix tests for Python 3.14 dev (#1218)
Tests were failing for two reasons:
- "asyncio.iscoroutinefunction()" is deprecated and must be replaced
with "inspect.iscoroutinefunction()"
- it seems the "PicklingError" now displays additional context at the
end of the exception message. I changed the implemntation of the tests
so that we check the presence of the expected string regardless of the
line it appears. It wasn't strictly required for the other non-failing
tests, but I generalized the implementation for consistency and
simplification.