Convert README from RST to MD (fix PyPI packaging)
The Readme was not compliant with PyPI renderer. Instead of using
reStructuredText, it was converted to GitHub Flavored Markdown (preserve
centered elements, notably).

To make it compatible with the HTML documentation generated by Sphinx,
we rely on the "myst_parser" extension. We probably should consider
migrating the whole documentation to Markdown.
There is also MkDocs, but it seems more intrusive in regards to the
existing Sphinx / ReadTheDocs configuration.